refactor(remote_control): 拆分端点与账本子模块保持单锁
将 manager.rs(5515 行,Wave 2 最大文件)由文件模块转为目录模块
manager/,按职责提取四个子模块,保持单 Arc<Mutex<Inner>> 与 pub 面不变:

- browse.rs: 主机文件浏览与会话产物分块读取(无状态纯函数)
- transfer.rs: 附件/会话上传下载缓冲(借 &mut Inner,不持锁)
- rpc.rs: RPC 幂等准入、响应整形、scope 校验与流/事件协议消息
- persistence.rs: 持久配置、RPC 账本、待撤销重放与自定义 Relay 设置

零 async 方法,parking_lot 同步锁;子模块作为 manager 后代可直接访问
父级私有项(Inner/类型/常量),无需改动任何可见性。facade 编排与
impl RemoteControlManager 留在 mod.rs。

验证: cargo fmt --check clean; fork-guard --fast 全过;
62/62 remote_control 测试通过;行为零变更。
